Move into amulet directory
