+++ MANIFEST.in
@@ -1,3 +1,3 @@
-recursive-include src/pybind11_extensions *.hpp *Config.cmake
+recursive-include src/amulet *.hpp *Config.cmake
 
 prune tests
