+++ pyproject.toml
@@ -48,7 +48,7 @@ readme = {file = ["README.md"], content-type = "text/markdown"}
 [tool.versioneer]
 VCS = "git"
 style = "pep440"
-versionfile_source = "src/pybind11_extensions/_version.py"
-versionfile_build = "pybind11_extensions/_version.py"
+versionfile_source = "src/amulet/pybind11/_version.py"
+versionfile_build = "amulet/pybind11/_version.py"
 tag_prefix = ""
-parentdir_prefix = "pybind11_extensions-"
+parentdir_prefix = "amulet_pybind11_extensions-"
