+++ tests/CMakeLists.txt
@@ -1,6 +1,6 @@
 cmake_minimum_required(VERSION 3.13)
 
-project(pybind11_extensions LANGUAGES CXX)
+project(amulet_pybind11_extensions_tests LANGUAGES CXX)
 
 # Set C++20
 set(CMAKE_CXX_STANDARD 20)
@@ -19,11 +19,11 @@ endif()
 
 # Add pybind11
 find_package(pybind11 CONFIG REQUIRED)
-find_package(pybind11_extensions CONFIG REQUIRED)
+find_package(amulet_pybind11_extensions CONFIG REQUIRED)
 
 function(add_text_ext name)
     pybind11_add_module(${name})
-    target_link_libraries(${name} PRIVATE pybind11_extensions)
+    target_link_libraries(${name} PRIVATE amulet_pybind11_extensions)
     target_compile_definitions(${name} PRIVATE PYBIND11_DETAILED_ERROR_MESSAGES)
     target_sources(${name} PRIVATE test_${name}.cpp)
     install(TARGETS ${name} DESTINATION ".")
